+++ scripts/helpers/animation/crit/styles/personaCrit.js
@@ -1,4 +1,4 @@
-import { getSetting } from "../../../misc.js";
+import { getSetting, MODULE_ID } from "../../../misc.js";
 /**
  *  Conversion method from css clip path
  * r = """0% 55%, 2% 52%, 9% 51%, 15% 44%, 23% 40%, 32% 38%, 34% 36%, 35% 35%, 41% 28%, 43% 30%, 50% 26%, 53% 27%, 58% 26%, 59% 26%, 62% 24%, 65% 25%, 71% 23%, 78% 15%, 85% 14%, 89% 14%, 95% 11%, 97% 12%, 100% 9%, 100% 55%, 97% 53%, 96% 55%, 92% 55%, 80% 56%, 72% 57%, 69% 58%, 64% 63%, 62% 63%, 61% 65%, 59% 63%, 57% 62%, 55% 64%, 53% 65%, 49% 63%, 43% 63%, 39% 64%, 37% 65%, 36% 65%, 34% 68%, 32% 67%, 29% 72%, 27% 71%, 27% 73%, 24% 72%, 22% 73%, 20% 70%, 16% 73%, 14% 71%, 13% 72%, 10% 71%, 5% 72%, 6% 70%, 0% 73%"""
@@ -95,7 +95,7 @@ export function personaCrit(actor, users, config) {
 
         video.onloadeddata = async () => {
             //await Sequencer.Preloader.preloadForClients([config.art, config.sfx]);
-            new Sequence()
+            new Sequence({moduleName: game.modules.get(MODULE_ID).title})
                 // BG Color
                 .effect()
                 .syncGroup(`p5-crit-${actor?.uuid}`)
@@ -144,11 +144,11 @@ export function personaCrit(actor, users, config) {
                 .volume(config.volume)
                 .forUsers(users)
                 .delay(config.delay)
-                .play();
+                .play({preload: true });
         };
     } else {
         //await Sequencer.Preloader.preloadForClients([config.art, config.sfx]);
-        new Sequence()
+        new Sequence({moduleName: game.modules.get(MODULE_ID).title})
             // BG Color
             .effect()
             .syncGroup(`p5-crit-${actor?.uuid}`)
@@ -210,6 +210,6 @@ export function personaCrit(actor, users, config) {
             .volume(config.volume)
             .forUsers(users)
             .delay(config.delay)
-            .play();
+            .play({preload: true });
     }
 }